image/requirements.txt
langchain==0.2.2
langchain-openai==0.1.8 
langchain-community
openai==1.31.1
chromadb==0.5.0
datasets==2.10.1 
pinecone-client==3.1.0 
tiktoken==0.7.0
python-dotenv==1.0.1
unstructured==0.14.4
boto3==1.34.117
mangum==0.17.0
uvicorn==0.30.0
streamlit==1.36.0
pypdf==4.2.0
python-pptx==0.6.23
python-docx==1.1.2